feat(crypto-block): store round on block (#376)
* ensure valid genesis block

* add round field to block

* update genesis block

* sync round to db

* style: resolve style guide violations

* change `numberOfTransactions` to uint16

* regenerate

* style: resolve style guide violations

* fix tests

* style: resolve style guide violations

---------

+++ packages/serializer/source/serializer.ts
@@ -134,7 +134,7 @@ export class Serializer implements Contracts.Serializer.Serializer {
 				for (const transaction of value) {
 					const serialized: Buffer = await this.transactionUtils.toBytes(transaction);
 
-					result.writeUint32(serialized.length);
+					result.writeUint16(serialized.length);
 					result.writeBytes(serialized);
 				}
 				continue;
@@ -242,7 +242,7 @@ export class Serializer implements Contracts.Serializer.Serializer {
 				target[property] = [];
 
 				for (let index = 0; index < (target as any).numberOfTransactions; index++) {
-					target[property].push(source.readBytes(source.readUint32()));
+					target[property].push(source.readBytes(source.readUint16()));
 				}
 				continue;
 			}
